pyproject.toml
[project]
name = "cz-telemetry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "pyqt6>=6.9.1",
    "pyserial>=3.5",
]

[tool]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = [
    "**/node_modules",
    "**/__pycache__",
    "**/*venv/**/*",
]

[tool.uv.workspace]
members = [
    "ui",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
